Deepen SAMS modules and remove undeployed read Lambdas.
Consolidate match loading and teams sync in lib/sams, enforce cache-peek SSR loaders, and delete unused read handlers.
